Fix #3
- Changed the offset of the sprite texture so that there are no overlaps with the neighbouring texture.
- Also update to 1.20.4

+++ src/main/java/me/wolfii/DrawContextFloatDrawTexture.java
@@ -3,5 +3,5 @@ package me.wolfii;
 import net.minecraft.util.Identifier;
 
 public interface DrawContextFloatDrawTexture {
-    default void centered_crosshair$drawTexture(Identifier texture, float x, float y, float width, float height) {}
+    default void centered_crosshair$drawTexture(Identifier texture, float x, float y, int width, int height) {}
 }
